[Refactor]:구조수정
